# api.yml
openapi: 3.0.0
info:
  title: ICETECH DNS API
  version: 1.0.0
paths:
  /domains/{domainName}:
    get:
      summary: Get domain details
      parameters:
        - name: domainName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Domain'
  /ipaddresses/{ip}:
    get:
      summary: Get IP address details
      parameters:
        - name: ip
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IPaddress'
  /whois/{query}:
    get:
      summary: Get WHOIS data for an IP range, domain, or AS
      parameters:
        - name: query
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WhoisData'
components:
  schemas:
    IPaddress:
      type: object
      properties:
        ip:
          type: string
        mask:
          type: string
        advertisers:
          type: array
          items:
            $ref: '#/components/schemas/ASNinfo'
        timestamp:
          type: string
          format: date-time
        other:
          type: string
        notes:
          type: string
        tags:
          type: object
          additionalProperties:
            type: string
    ASNinfo:
      type: object
      properties:
        asn:
          type: integer
        country:
          type: string
        registry:
          type: string
        timestamp:
          type: string
          format: date-time
        other:
          type: string
        notes:
          type: string
        tags:
          type: object
          additionalProperties:
            type: string
    DNSEntry:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        value:
          type: string
        ttl:
          type: integer
        timestamp:
          type: string
          format: date-time
        other:
          type: string
        notes:
          type: string
        tags:
          type: object
          additionalProperties:
            type: string
    Domain:
      type: object
      properties:
        name:
          type: string
        records:
          type: array
          items:
            $ref: '#/components/schemas/DNSEntry'
        parent:
          type: string
        paths:
          type: array
          items:
            $ref: '#/components/schemas/Path'
        addresses:
          type: array
          items:
            $ref: '#/components/schemas/IPaddress'
        timestamp:
          type: string
          format: date-time
        other:
          type: string
        notes:
          type: string
        tags:
          type: object
          additionalProperties:
            type: string
    Path:
      type: object
      properties:
        path:
          type: string
        other:
          type: string
        notes:
          type: string
        tags:
          type: object
          additionalProperties:
            type: string
    Nameserver:
      type: object
      properties:
        name:
          type: string
        ip:
          $ref: '#/components/schemas/IPaddress'
        timestamp:
          type: string
          format: date-time
        other:
          type: string
        notes:
          type: string
        tags:
          type: object
          additionalProperties:
            type: string
    RegistrarData:
      type: object
      properties:
        name:
          type: string
        url:
          type: string
        countryCode:
          type: string
        phone:
          type: string
        fax:
          type: string
        address:
          type: array
          items:
            $ref: '#/components/schemas/IPaddress'
        timestamp:
          type: string
          format: date-time
        other:
          type: string
        notes:
          type: string
        tags:
          type: object
          additionalProperties:
            type: string
    WhoisData:
      type: object
      properties:
        ip_range:
          $ref: '#/components/schemas/IPaddress'
        domain_name:
          type: string
        asn:
          type: integer
        country:
          type: string
        created:
          type: string
          format: date-time
        updated:
          type: string
          format: date-time
        registrar:
          type: string
        timestamp:
          type: string
          format: date-time
        other:
          type: string
        notes:
          type: string
        nameservers:
          type: array
          items:
            $ref: '#/components/schemas/Nameserver'