# internal/api/gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.4.1 DO NOT EDIT.
package api

import (
	"fmt"
	"net/http"
	"time"

	"github.com/gorilla/mux"
	"github.com/oapi-codegen/runtime"
)

// ASNinfo defines model for ASNinfo.
type ASNinfo struct {
	Asn       *int               `json:"asn,omitempty"`
	Country   *string            `json:"country,omitempty"`
	Notes     *string            `json:"notes,omitempty"`
	Other     *string            `json:"other,omitempty"`
	Registry  *string            `json:"registry,omitempty"`
	Tags      *map[string]string `json:"tags,omitempty"`
	Timestamp *time.Time         `json:"timestamp,omitempty"`
}

// DNSEntry defines model for DNSEntry.
type DNSEntry struct {
	Name      *string            `json:"name,omitempty"`
	Notes     *string            `json:"notes,omitempty"`
	Other     *string            `json:"other,omitempty"`
	Tags      *map[string]string `json:"tags,omitempty"`
	Timestamp *time.Time         `json:"timestamp,omitempty"`
	Ttl       *int               `json:"ttl,omitempty"`
	Type      *string            `json:"type,omitempty"`
	Value     *string            `json:"value,omitempty"`
}

// Domain defines model for Domain.
type Domain struct {
	Addresses *[]IPaddress       `json:"addresses,omitempty"`
	Name      *string            `json:"name,omitempty"`
	Notes     *string            `json:"notes,omitempty"`
	Other     *string            `json:"other,omitempty"`
	Parent    *string            `json:"parent,omitempty"`
	Paths     *[]Path            `json:"paths,omitempty"`
	Records   *[]DNSEntry        `json:"records,omitempty"`
	Tags      *map[string]string `json:"tags,omitempty"`
	Timestamp *time.Time         `json:"timestamp,omitempty"`
}

// IPaddress defines model for IPaddress.
type IPaddress struct {
	Advertisers *[]ASNinfo         `json:"advertisers,omitempty"`
	Ip          *string            `json:"ip,omitempty"`
	Mask        *string            `json:"mask,omitempty"`
	Notes       *string            `json:"notes,omitempty"`
	Other       *string            `json:"other,omitempty"`
	Tags        *map[string]string `json:"tags,omitempty"`
	Timestamp   *time.Time         `json:"timestamp,omitempty"`
}

// Nameserver defines model for Nameserver.
type Nameserver struct {
	Ip        *IPaddress         `json:"ip,omitempty"`
	Name      *string            `json:"name,omitempty"`
	Notes     *string            `json:"notes,omitempty"`
	Other     *string            `json:"other,omitempty"`
	Tags      *map[string]string `json:"tags,omitempty"`
	Timestamp *time.Time         `json:"timestamp,omitempty"`
}

// Path defines model for Path.
type Path struct {
	Notes *string            `json:"notes,omitempty"`
	Other *string            `json:"other,omitempty"`
	Path  *string            `json:"path,omitempty"`
	Tags  *map[string]string `json:"tags,omitempty"`
}

// WhoisData defines model for WhoisData.
type WhoisData struct {
	Asn         *int          `json:"asn,omitempty"`
	Country     *string       `json:"country,omitempty"`
	Created     *time.Time    `json:"created,omitempty"`
	DomainName  *string       `json:"domain_name,omitempty"`
	IpRange     *IPaddress    `json:"ip_range,omitempty"`
	Nameservers *[]Nameserver `json:"nameservers,omitempty"`
	Notes       *string       `json:"notes,omitempty"`
	Other       *string       `json:"other,omitempty"`
	Registrar   *string       `json:"registrar,omitempty"`
	Timestamp   *time.Time    `json:"timestamp,omitempty"`
	Updated     *time.Time    `json:"updated,omitempty"`
}

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Get domain details
	// (GET /domains/{domainName})
	GetDomainsDomainName(w http.ResponseWriter, r *http.Request, domainName string)
	// Get IP address details
	// (GET /ipaddresses/{ip})
	GetIpaddressesIp(w http.ResponseWriter, r *http.Request, ip string)
	// Get WHOIS data for an IP range, domain, or AS
	// (GET /whois/{query})
	GetWhoisQuery(w http.ResponseWriter, r *http.Request, query string)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// GetDomainsDomainName operation middleware
func (siw *ServerInterfaceWrapper) GetDomainsDomainName(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "domainName" -------------
	var domainName string

	err = runtime.BindStyledParameterWithOptions("simple", "domainName", mux.Vars(r)["domainName"], &domainName, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "domainName", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetDomainsDomainName(w, r, domainName)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetIpaddressesIp operation middleware
func (siw *ServerInterfaceWrapper) GetIpaddressesIp(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "ip" -------------
	var ip string

	err = runtime.BindStyledParameterWithOptions("simple", "ip", mux.Vars(r)["ip"], &ip, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "ip", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetIpaddressesIp(w, r, ip)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetWhoisQuery operation middleware
func (siw *ServerInterfaceWrapper) GetWhoisQuery(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "query" -------------
	var query string

	err = runtime.BindStyledParameterWithOptions("simple", "query", mux.Vars(r)["query"], &query, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "query", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetWhoisQuery(w, r, query)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, GorillaServerOptions{})
}

type GorillaServerOptions struct {
	BaseURL          string
	BaseRouter       *mux.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r *mux.Router) http.Handler {
	return HandlerWithOptions(si, GorillaServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r *mux.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, GorillaServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options GorillaServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = mux.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.HandleFunc(options.BaseURL+"/domains/{domainName}", wrapper.GetDomainsDomainName).Methods("GET")

	r.HandleFunc(options.BaseURL+"/ipaddresses/{ip}", wrapper.GetIpaddressesIp).Methods("GET")

	r.HandleFunc(options.BaseURL+"/whois/{query}", wrapper.GetWhoisQuery).Methods("GET")

	return r
}